# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager API",
        "contact": {
            "email": "wesleymuniz20@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://mit-license.org"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Task Endpoint"
                ],
                "summary": "Get list of tasks",
                "description": "Returns list paginated of tasks",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Query to search tasks by title",
                        "schema": {
                            "type": "string",
                            "example": "title"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Query to sort by creation: asc|desc",
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "asc": {
                                "summary": "An asc sort --default",
                                "value": "asc"
                            },
                            "desc": {
                                "summary": "An desc sort",
                                "value": "desc"
                            }
                        }
                    },
                    {
                        "name": "is_done",
                        "in": "query",
                        "description": "Query to filter completed tasks",
                        "schema": {
                            "type": "boolean",
                            "example": "false"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskCollection"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Task Endpoint"
                ],
                "summary": "Post a new Task",
                "description": "Returns id of the created task",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Task created successfully!"
                                                },
                                                "data": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "format": "uuid"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Task Endpoint"
                ],
                "summary": "Get task info by id",
                "description": "Returns task data by id",
                "operationId": "90b3cd4d2d4b416dc989d2055dbed06c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Path to find task by id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "2a825edc-95f6-49a9-b46a-d2f296e7e1ce"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Task Endpoint"
                ],
                "summary": "Put task by id",
                "description": "Returns id of the updated task",
                "operationId": "1aaf37e74360ef4c10fadaee02ae77fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Path to find task by id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "2a825edc-95f6-49a9-b46a-d2f296e7e1ce"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Task updated successfully!"
                                                },
                                                "data": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "format": "uuid"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Task Endpoint"
                ],
                "summary": "Delete task by id",
                "operationId": "ad582480cd57c2a4e3b6a426943eeca5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Path to find task by id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "2a825edc-95f6-49a9-b46a-d2f296e7e1ce"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TaskRequest": {
                "title": "TaskRequest",
                "properties": {
                    "project_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 3
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "is_done": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
            "TaskCollection": {
                "title": "TaskCollection",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Task"
                        }
                    }
                },
                "type": "object"
            },
            "TaskResource": {
                "title": "TaskResource",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Task"
                    }
                },
                "type": "object"
            },
...
            "Task": {
                "title": "Task",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "project_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "is_done": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}